+++ src/Telegram.Bot.Extensions.LoginWidget/ButtonStyle.cs
@@ -1,0 +1,9 @@
+﻿namespace Telegram.Bot.Extensions.LoginWidget
+{
+    public enum ButtonStyle
+    {
+        Large,
+        Medium,
+        Small
+    }
+}