+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### Added
 
+- A 'WidgetEmbedCodeGenerator' class
 - Tests for the 'LoginWidget.CheckAuthorization' method
 - Login Widget with support for validating authorization hashes